+++ core/src/test/kotlin/sift/core/api/testdata/set2/Interfaces.kt
@@ -1,12 +1,12 @@
 package sift.core.api.testdata.set2
 
-import sift.core.api.testdata.set1.Payload
-
 object Interfaces {
     interface A
     interface B : A
     interface C : B
     interface D : C
+
+    interface E : B, C, D
 
     abstract class Base
 
@@ -14,4 +14,9 @@ object Interfaces {
     open class ImplementorB : Base(), A
     open class ImplementorC : ImplementorB()
     open class ImplementorD : ImplementorC(), D
+
+    open class ImplAx : A
+    open class ImplBx : B, ImplAx()
+    open class ImplCx : C, ImplBx()
+    open class ImplDx : D, ImplCx()
 }